More changes to national Ideas
Buffed US national ideas to be more competitive with vermont and texas. Buffed england land and naval ideas so it has a chance to survive without forming GB and to discourage england players to forming ireland for land warfare. Buffed GB land ideas since GB was also a land power and 5% discipline alone is not enough to survive the french.